+++ dm.js
@@ -1,5 +1,5 @@
 /*
- * DOM Markers 0.2.3
+ * DOM Markers 0.3.0
  * Copyright 2013 Eugene Poltorakov
  * Licensed under the MIT License: http://www.opensource.org/licenses/mit-license.php
  */
@@ -22,31 +22,44 @@ DMUtils = {
       }
     }
   },
+  /**
+   * Map utility function
+   * @param {Array} arr
+   * @param {Function} callback
+   * @param {*?} context
+   * @returns {*}
+   */
   map : function(arr, callback, context) {
     return arr.map(callback, context);
   },
   /**
    * @param {String} selector
-   * @param {HTMLElement|HTMLDocument|?} ctx
+   * @param {Element|HTMLDocument|?} ctx
    * @returns {NodeList}
    */
   all : function(selector, ctx) {
-    if (!(ctx instanceof HTMLElement || ctx instanceof HTMLDocument)) {
+    if (!(ctx instanceof Element || ctx instanceof HTMLDocument)) {
       ctx = document;
     }
 
     return ctx.querySelectorAll(selector);
   },
+  /**
+   * Trim
+   * @param {string} str
+   * @returns {string}
+   */
   trim : function(str) {
     return str.trim();
   },
   /**
    *
-   * @param {HTMLElement} node
+   * @param {Element} node
+   * @param {string} attrName
    * @return {Array.<{name:String,args:Array}>}
    */
-  getModules : function(node) {
-    return DMUtils.map(node.getAttribute('data-marker').match(/([a-z\-]+(\[[^[]+\])?)/ig) || [], function(str) {
+  getModules : function(node, attrName) {
+    return DMUtils.map(node.getAttribute(attrName).match(/([a-z\-]+(\[[^[]+\])?)/ig) || [], function(str) {
       var parts = str.match(/[^\[\]]+/ig),
         name,
         args;
@@ -77,9 +90,22 @@ DMUtils = {
       };
     });
   },
+  /**
+   * Filter utility function
+   * @param arr
+   * @param callback
+   * @returns {*}
+   */
   filter : function(arr, callback) {
     return arr.filter(callback);
   },
+  /**
+   * Filter modules (exclude already executed on node)
+   * @param {Element} node
+   * @param {Array} list
+   * @param {Object} modules
+   * @returns {*}
+   */
   filterModules : function(node, list, modules) {
     return DMUtils.filter(list, function(module) {
       var _data = node._dm || (node._dm = {}),
@@ -96,6 +122,7 @@ DMUtils = {
     });
   },
   /**
+   * Return new unique id
    * @returns {Function}
    */
   uuid : (function() {
@@ -107,12 +134,12 @@ DMUtils = {
 };
 
 /**
- *
+ * @param {string} name
  * @param {Function?} callback
  * @param {*?} context
  * @constructor
  */
-function DMModule(callback, context) {
+function DMModule(name, callback, context) {
   this.uuid = DMUtils.uuid();
 
   this._before = [];
@@ -121,15 +148,22 @@ function DMModule(callback, context) {
     callback : callback,
     context  : context
   };
+  this.name = name;
   this.ready = false;
 }
 
+/**
+ * Sort module
+ * @param {Array} arr
+ * @returns {this}
+ * @private
+ */
 DMModule.prototype._sort = function(arr){
   arr.sort(function(a, b){
     var result;
 
     if (a.weight === b.weight) {
-      result = 0;
+      result = 0;//todo - check that the native sort works correctly when arr.length > 10
     }
     else {
       result = a.weight > b.weight ? 1 : -1;
@@ -139,6 +173,12 @@ DMModule.prototype._sort = function(arr){
   });
   return this;
 };
+
+/**
+ * Preparation method
+ * - Sort the callbacks
+ * @returns {this}
+ */
 DMModule.prototype.prepare = function(){
   this._sort(this._before)._sort(this._after);
   this.ready = true;
@@ -185,6 +225,14 @@ DMModule.prototype.after = function(callback, context, weight){
   return id;
 };
 
+/**
+ * DMExec
+ * Execution constructor; Manage callbacks execution
+ * @param module
+ * @param node
+ * @param args
+ * @constructor
+ */
 function DMExec(module, node, args) {
   this.module = module;
   this.node = node;
@@ -197,6 +245,10 @@ function DMExec(module, node, args) {
   this._timer = null;
 }
 
+/**
+ * Execution states
+ * @type {{INITIAL: number, BEFORE: number, MAIN: number, AFTER: number, FINISHED: number}}
+ */
 DMExec.STATES = {
   INITIAL  : 0,
   BEFORE   : 1,
@@ -205,11 +257,18 @@ DMExec.STATES = {
   FINISHED : 4
 };
 
+/**
+ * Execution types
+ * @type {{NEXT: string, STOP: string}}
+ */
 DMExec.TYPES = {
   NEXT : 'next',
   STOP : 'stop'
 };
 
+/**
+ * Force current execution manager instance to execute next callback
+ */
 DMExec.prototype.next = function(){
   if (this._waiting) {
     this._waiting = false;
@@ -217,7 +276,11 @@ DMExec.prototype.next = function(){
   }
   this._index++;
   this.execute(DMExec.TYPES.NEXT);
+  //todo - should we stop any other code below the next call ?
 };
+/**
+ * Stops current execution
+ */
 DMExec.prototype.stop = function(){
   if (this._waiting) {
     this._waiting = false;
@@ -228,9 +291,9 @@ DMExec.prototype.stop = function(){
   this.execute(DMExec.TYPES.STOP);
 };
 /**
- *
+ * Proceed callback execution
  * @param {String?} type
- * @returns {*}
+ * @returns {DMExec}
  */
 DMExec.prototype.execute = function(type){
   var states = DMExec.STATES,
@@ -287,6 +350,12 @@ DMExec.prototype.execute = function(type){
 
   return this;
 };
+
+/**
+ * Initiate timeout
+ * @param {number?} timeout - wait timeout in ms; default: 5000
+ * @param {boolean?} stop - will abort execution if timeout reached & value is true; default: false
+ */
 DMExec.prototype.wait = function(timeout, stop){
   var self = this;
 
@@ -297,7 +366,40 @@ DMExec.prototype.wait = function(timeout, stop){
   }, timeout || 5000);
 };
 
+/**
+ * @returns {*}
+ */
+DMExec.prototype.children = function() {
+  //todo - should accept role
+  //todo - should cache
+  //todo - cover with tests
+  var attrName,
+    nodes,
+    result = {};
 
+  attrName = 'data-' + this.module.name;
+  nodes = DMUtils.all('[' + attrName + ']', this.node);
+
+  DMUtils.each(Array.prototype.slice.call(nodes), function(node) {
+    DMUtils.each(DMUtils.getModules(node, attrName), function(module) {
+      if (!result[module.name]) {
+        result[module.name] = [];
+      }
+      result[module.name].push({
+        node : node,
+        args : module.args
+      });
+    });
+  }, this);
+
+  return result;
+};
+
+/**
+ * Singleton; Main library wrapper;
+ *
+ * {{add:Function,before:Function,after:Function,go:Function,detach:Function,remove:Function,removeAll:Function}}
+ */
 DM = (function(options) {
   var _modules = {},
     _engine;
@@ -308,8 +410,7 @@ DM = (function(options) {
         //_engine = options.engines.y;
         options.engines.y().use('node-base', 'array-extras', function(Y) {
           DMUtils.all = function(selector, ctx){
-            //todo - add context support
-            return Y.all(selector).getDOMNodes();
+            return (ctx ? Y.one(ctx) : Y).all(selector).getDOMNodes();
           };
 
           DMUtils.map = function(arr, callback, context){
@@ -367,7 +468,7 @@ DM = (function(options) {
    * @returns {DMModule}
    */
   function createModule(name, callback, context) {
-    return _modules[name] = new DMModule(callback, context);
+    return _modules[name] = new DMModule(name, callback, context);
   }
 
   /**
@@ -440,12 +541,17 @@ DM = (function(options) {
 
       return module.after(callback, context, weight);
     },
+    /**
+     * Initiate callbacks execution
+     * @returns {DM}
+     */
     go : function() {
+      //todo - should accept & execute only asked module(s): Array.<string>
       initEngine(function() {
         var nodes = DMUtils.all('[data-marker]', options.env.document);
 
         DMUtils.each(Array.prototype.slice.call(nodes), function(node) {
-          var modules = DMUtils.getModules(node);
+          var modules = DMUtils.getModules(node, 'data-marker');
 
           modules = DMUtils.filterModules(node, modules, _modules);
 
@@ -537,7 +643,7 @@ DM = (function(options) {
     }
   };
 })({
-  env     : {
+  env : {
     win      : typeof window !== 'undefined' && window,
     document : typeof document !== 'undefined' && document
   },