+++ tests/unit/assets/dm-test.js
@@ -246,7 +246,7 @@ YUI.add('dm-test', function (Y) {
 
     _should : {
       error : {
-        "should throw an error when trying to redefine module existing body" : 'Module(module) main callback is already defined'
+        "Should throw an error when trying to redefine module existing body" : 'Module(module) main callback is already defined'
       }
     },
 
@@ -1033,8 +1033,250 @@ YUI.add('dm-test', function (Y) {
     }
   }));
 
-  //todo - test wait with different parameters
-  //todo - test that wait isn't stop other process (other modules or elements)
-  //todo - test concurrent modules execution
-  //todo - test next
-}, '0.2.3', {requires:['dm', 'test']});
+  Y.Test.Runner.add(new Y.Test.Case({
+    name: 'DOM Markers : Testing DMExec.wait()',
+    setUp : function() {
+      Y.one('#dump').setHTML('<div id="node-a" data-marker="foo"></div>');
+
+      this.nodes = {
+        a : Y.one('#node-a')
+      };
+    },
+    tearDown : function() {
+      Y.one('#dump').empty();
+      DM.removeAll();
+    },
+    "Should continue after timeout" : function() {
+      var test = this;
+
+      DM.add('foo', function() {
+        this.wait(200, false);
+      });
+
+      DM.after('foo', function() {
+        test.resume();
+      });
+
+      DM.go();
+
+      this.wait(500);
+    },
+    "Should stop after timeout" : function() {
+      DM.add('foo', function() {
+        this.wait(100, true);
+      });
+
+      DM.after('foo', function() {
+        throw new Error('foo:after should not be called');
+      });
+
+      DM.go();
+
+      this.wait(function() {
+        //everything is ok
+      }, 200);
+    },
+    "Should wait 200 ms in 'before' section and makes sure that timeout is nearly accurate" : function() {
+      var start,
+        delta;
+
+      DM.before('foo', function() {
+        this.wait(200, false);
+      });
+
+      DM.after('foo', function() {
+        delta = Date.now() - start;
+      });
+
+      start = Date.now();
+
+      DM.go();
+
+      this.wait(function() {
+        Assert.isNotUndefined(delta, 'Delta should not be undefined');
+        if (Math.abs(delta - 200) > 2) {
+          throw new Error(Y.Lang.sub("Timeout is not accurate. Expected : 200. Actual: {value}", {
+            value : delta
+          }));
+        }
+      }, 300);
+    },
+    "Should wait 200 ms in 'body' section and makes sure that timeout is nearly accurate" : function() {
+      var start,
+        delta;
+
+      DM.add('foo', function() {
+        this.wait(200, false);
+      });
+
+      DM.after('foo', function() {
+        delta = Date.now() - start;
+      });
+
+      start = Date.now();
+
+      DM.go();
+
+      this.wait(function() {
+        Assert.isNotUndefined(delta, 'Delta should not be undefined');
+        if (Math.abs(delta - 200) > 2) {
+          throw new Error(Y.Lang.sub("Timeout is not accurate. Expected : 200. Actual: {value}", {
+            value : delta
+          }));
+        }
+      }, 300);
+    },
+    "Should wait 200 ms in 'after' section and makes sure that timeout is nearly accurate" : function() {
+      var start,
+        delta;
+
+      DM.after('foo', function() {
+        this.wait(200, false);
+      });
+
+      DM.after('foo', function() {
+        delta = Date.now() - start;
+      });
+
+      start = Date.now();
+
+      DM.go();
+
+      this.wait(function() {
+        Assert.isNotUndefined(delta, 'Delta should not be undefined');
+        if (Math.abs(delta - 200) > 2) {
+          throw new Error(Y.Lang.sub("Timeout is not accurate. Expected : 200. Actual: {value}", {
+            value : delta
+          }));
+        }
+      }, 300);
+    },
+    "Should check that `wait` without second parameter will continue the execution" : function() {
+      var test = this;
+
+      DM.add('foo', function() {
+        this.wait(10);
+      });
+
+
+      DM.after('foo', function() {
+        test.resume();
+      });
+
+      DM.go();
+
+      this.wait(100);
+    }
+  }));
+
+  Y.Test.Runner.add(new Y.Test.Case({
+    name: 'DOM Markers : Testing DMExec.next',
+    setUp : function() {
+      Y.one('#dump').setHTML('<div id="node-a" data-marker="foo"></div>');
+
+      this.nodes = {
+        a : Y.one('#node-a')
+      };
+    },
+    tearDown : function() {
+      Y.one('#dump').empty();
+      DM.removeAll();
+    },
+    "Should continue execution by calling next() before stop timeout is fired" : function() {
+      var mock = new Y.Mock,
+        callback;
+
+      callback = function() {
+        var exec = this;
+
+        Y.later(25, this, function() {
+          this.next();
+        });
+
+        //will stop if next will not be called
+        this.wait(50, true);
+      };
+
+      Y.Mock.expect(mock, {
+        method : 'before',
+        run    : callback
+      });
+
+      Y.Mock.expect(mock, {
+        method : 'body',
+        run    : callback
+      });
+
+      Y.Mock.expect(mock, {
+        method : 'after',
+        run    : callback
+      });
+
+      DM.before('foo', mock.before);
+      DM.add('foo', mock.body);
+      DM.after('foo', mock.after);
+
+      DM.go();
+
+      this.wait(function() {
+        Y.Mock.verify(mock);
+      }, 300);
+    }
+  }));
+
+  Y.Test.Runner.add(new Y.Test.Case({
+    name: 'DOM Markers : Test concurrent modules execution',
+    setUp : function() {
+      Y.one('#dump').setHTML('<div id="node-a" data-marker="foo,bar"></div>');
+
+      this.nodes = {
+        a : Y.one('#node-a')
+      };
+    },
+    tearDown : function() {
+      Y.one('#dump').empty();
+      DM.removeAll();
+    },
+    "Should execute foo:body (with DMExec.wait) > bar:body > foo:after" : function() {
+      var mock = new Y.Mock,
+        test = this,
+        log = [];
+
+      Y.Mock.expect(mock, {
+        method : 'fooBody',
+        run : function() {
+          log.push('foo:body');
+          this.wait(50);
+        }
+      });
+
+      Y.Mock.expect(mock, {
+        method : 'fooAfter',
+        run : function() {
+          log.push('foo:after');
+        }
+      });
+
+      Y.Mock.expect(mock, {
+        method : 'barBody',
+        run : function() {
+          log.push('bar:body');
+        }
+      });
+
+      DM.add('foo', mock.fooBody);
+
+      DM.after('foo', mock.fooAfter);
+
+      DM.add('bar', mock.barBody);
+
+      DM.go();
+
+      this.wait(function() {
+        Assert.areSame('foo:body,bar:body,foo:after', log.join(','), 'Execution order is not correct');
+
+        Y.Mock.verify(mock);
+      }, 300);
+    }
+  }));
+}, '0.2.4', {requires:['dm', 'test']});